+++ packages/rum/src/domain/record/observers/mediaInteractionObserver.ts
@@ -5,8 +5,8 @@ import type { MediaInteraction } from '../../../types'
 import { MediaInteractionType } from '../../../types'
 import { getNodePrivacyLevel } from '../privacy'
 import { getSerializedNodeId, hasSerializedNode } from '../serializationUtils'
-import type { ListenerHandler } from '../utils'
-import { getEventTarget } from '../utils'
+import type { ListenerHandler } from './utils'
+import { getEventTarget } from './utils'
 
 export type MediaInteractionCallback = (p: MediaInteraction) => void
