+++ packages/rum/src/domain/record/observers/moveObserver.ts
@@ -1,10 +1,11 @@
 import { addEventListeners, addTelemetryDebug, DOM_EVENT, throttle } from '@datadog/browser-core'
 import { getSerializedNodeId, hasSerializedNode } from '../serializationUtils'
-import type { ListenerHandler } from '../utils'
-import { getEventTarget, isTouchEvent } from '../utils'
+import { isTouchEvent } from '../utils'
 import type { MousePosition } from '../../../types'
 import { IncrementalSource } from '../../../types'
 import { convertMouseEventToLayoutCoordinates } from '../viewports'
+import type { ListenerHandler } from './utils'
+import { getEventTarget } from './utils'
 
 const MOUSE_MOVE_OBSERVER_THRESHOLD = 50
