+++ packages/rum/src/domain/record/observers/observers.ts
@@ -1,5 +1,4 @@
 import type { LifeCycle, RumConfiguration } from '@datadog/browser-rum-core'
-import type { ListenerHandler } from '../utils'
 import type { ElementsScrollPositions } from '../elementsScrollPositions'
 import type { ShadowRootsController } from '../shadowRootsController'
 import type { MousemoveCallBack } from './moveObserver'
@@ -22,6 +21,7 @@ import type { MutationCallBack } from './mutationObserver'
 import { initMutationObserver } from './mutationObserver'
 import type { FocusCallback } from './focusObserver'
 import { initFocusObserver } from './focusObserver'
+import type { ListenerHandler } from './utils'
 
 interface ObserverParam {
   lifeCycle: LifeCycle
