+++ packages/rum/src/domain/record/observers/viewportResizeObserver.ts
@@ -1,8 +1,8 @@
 import { throttle, DOM_EVENT, addEventListeners, noop } from '@datadog/browser-core'
 import { initViewportObservable } from '@datadog/browser-rum-core'
 import type { ViewportResizeDimension, VisualViewportRecord } from '../../../types'
-import type { ListenerHandler } from '../utils'
 import { getVisualViewport } from '../viewports'
+import type { ListenerHandler } from './utils'
 
 const VISUAL_VIEWPORT_OBSERVER_THRESHOLD = 200
