+++ packages/rum/src/domain/record/utils.ts
@@ -1,17 +1,9 @@
 import { assign, timeStampNow } from '@datadog/browser-core'
-import { isNodeShadowHost } from '@datadog/browser-rum-core'
 import type { BrowserIncrementalData, BrowserIncrementalSnapshotRecord } from '../../types'
 import { RecordType } from '../../types'
 
 export function isTouchEvent(event: MouseEvent | TouchEvent): event is TouchEvent {
   return Boolean((event as TouchEvent).changedTouches)
-}
-
-export function forEach<List extends { [index: number]: any }>(
-  list: List,
-  callback: (value: List[number], index: number, parent: List) => void
-) {
-  Array.prototype.forEach.call(list, callback as any)
 }
 
 export function assembleIncrementalSnapshot<Data extends BrowserIncrementalData>(
@@ -49,23 +41,4 @@ export function getPathToNestedCSSRule(rule: CSSRule): number[] | undefined {
   path.unshift(index)
 
   return path
-}
-
-export type ListenerHandler = () => void
-
-export function getEventTarget(event: Event): Node {
-  if (event.composed === true && isNodeShadowHost(event.target as Node)) {
-    return event.composedPath()[0] as Node
-  }
-  return event.target as Node
-}
-
-const recordIds = new WeakMap<Event, number>()
-let nextId = 1
-
-export function getRecordIdForEvent(event: Event): number {
-  if (!recordIds.has(event)) {
-    recordIds.set(event, nextId++)
-  }
-  return recordIds.get(event)!
 }
